+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Send HTTP status code 404 when attempting to access the file server while it is disabled
+
 ### Changed
 
 ### Removed